+++ service/neptunedata/doc.go
@@ -5,9 +5,9 @@
 //
 // The Amazon Neptune data API provides SDK support for more than 40 of Neptune's
 // data operations, including data loading, query execution, data inquiry, and
-// machine learning. It supports all three Neptune query languages (Gremlin,
-// openCypher and SPARQL), and is available in all SDK languages. It automatically
-// signs API requests and greatly simplifies integrating Neptune into your applications.
+// machine learning. It supports the Gremlin and openCypher query languages,
+// and is available in all SDK languages. It automatically signs API requests
+// and greatly simplifies integrating Neptune into your applications.
 //
 // See https://docs.aws.amazon.com/goto/WebAPI/neptunedata-2023-08-01 for more information on this service.
 //